상세페이지,게시글리스트 일부 올림

+++ src/main/java/com/exam/gagi/model/ItemImage.java
@@ -1,24 +1,24 @@
 package com.exam.gagi.model;
 
 public class ItemImage extends BaseEntity {
-    private Long id; // 이미지 ID
-    private Long itemId; // 상품 ID (FK)
+    private int id; // 이미지 ID
+    private int itemId; // 상품 ID (FK)
     private String imageUrl; // 이미지 경로
     private int sortOrder; // 정렬 순서
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getItemId() {
+    public int getItemId() {
         return itemId;
     }
 
-    public void setItemId(Long itemId) {
+    public void setItemId(int itemId) {
         this.itemId = itemId;
     }
