+++ src/main/java/com/exam/gagi/model/OrderDetailDto.java
@@ -4,16 +4,16 @@ import java.sql.Timestamp;
 import java.util.List;
 
 public class OrderDetailDto {
-    private Long orderId; // 주문 구분을 위한 ID
+    private int orderId; // 주문 구분을 위한 ID
     private Timestamp orderDate;
     private String shippingStatus;
     private List<OrderItemDto> orderItems;
 
-    public Long getOrderId() {
+    public int getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Long orderId) {
+    public void setOrderId(int orderId) {
         this.orderId = orderId;
     }
