+++ src/main/java/com/exam/gagi/model/OrderItem.java
@@ -4,36 +4,36 @@ import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 public class OrderItem {
-    private Long id; // 주문 상세 번호
-    private Long orderId; // 주문 ID (FK)
-    private Long itemId; // 상품 ID (FK)
+    private int id; // 주문 상세 번호
+    private int orderId; // 주문 ID (FK)
+    private int itemId; // 상품 ID (FK)
     private int quantity; // 수량
     private BigDecimal price; // 주문 당시 가격
     private String orderStatus; // 주문 상품 상태
     private Item item; // 상품 정보
     private Timestamp createdAt; // 생성일
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getOrderId() {
+    public int getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Long orderId) {
+    public void setOrderId(int orderId) {
         this.orderId = orderId;
     }
 
-    public Long getItemId() {
+    public int getItemId() {
         return itemId;
     }
 
-    public void setItemId(Long itemId) {
+    public void setItemId(int itemId) {
         this.itemId = itemId;
     }
